requirements/production.txt
Django==1.10.5
dj-database-url==0.4.2
djangorestframework==3.5.3
prettyconf==1.2.3
django-mptt==0.8.7
# psycopg2
factory-boy==2.8.1
requests==2.12.4
djangorestframework==3.5.3